+++ coffee-kiosk-solution.Data/ViewModels/DiscountViewModel.cs
@@ -9,7 +9,7 @@ namespace coffee_kiosk_solution.Data.ViewModels
     public class DiscountViewModel
     {
         public Guid Id { get; set; }
-        public float DiscountPercentage { get; set; }
+        public double DiscountPercentage { get; set; }
         public Guid ProductId { get; set; }
         public Guid CampaignId { get; set; }
         public int Status { get; set; }
